+++ Sources/SwiftyHaru/DrawingContext.swift
@@ -616,6 +616,35 @@ public final class DrawingContext {
         
         return lines.map { HPDF_Page_TextWidth(_page, $0) }.max()!
     }
+
+    /// Calculates the number of UTF8 characters which can be included within the specified width.
+    ///
+    /// - Parameters:
+    ///   - text:     The text to use for calculation.
+    ///   - width:    The width of the area to put the text.
+    ///   - wordwrap: When there are three words of "ABCDE", "FGH", and "IJKL", and the substring until "J" can be
+    ///               included within the width, if `wordwrap` parameter is `false` it returns 12, and if `wordwrap`
+    ///               parameter is `true`, it returns 10 (the end of the previous word).
+    /// - Returns:
+    ///   - `utf8Length`: The byte length which can be included within the specified width in current font size,
+    ///                   character spacing and word spacing.
+    ///   - `realWidth`:  The real width of the text.
+    public func measureText(_ text: String,
+                            width: Float,
+                            wordwrap: Bool) throws -> (utf8Length: Int, realWidth: Float) {
+
+        _setFontIfNeeded()
+
+        var realWidth: Float = 0
+        let utf8Length = HPDF_Page_MeasureText(_page, text, width, wordwrap ? HPDF_TRUE : HPDF_FALSE, &realWidth)
+
+        if HPDF_GetError(_documentHandle) != UInt(HPDF_OK) {
+            HPDF_ResetError(_documentHandle)
+            throw _document._error
+        }
+
+        return (utf8Length: Int(utf8Length), realWidth: realWidth)
+    }
     
     /// Gets the bounding box of the text in the current font size and leading. Text can be multiline.
     ///
@@ -688,6 +717,61 @@ public final class DrawingContext {
         }
         set {
             HPDF_Page_SetTextRenderingMode(_page, HPDF_TextRenderingMode(newValue.rawValue))
+        }
+    }
+
+    /// The default character spacing.
+    public static let defaultCharacterSpacing = Float(HPDF_DEF_CHARSPACE)
+
+    /// The minimum character spacing.
+    public static let minCharacterSpacing = Float(HPDF_MIN_CHARSPACE)
+
+    /// The maximum character spacing.
+    public static let maxCharacterSpacing = Float(HPDF_MAX_CHARSPACE)
+
+    /// The current value of the page's character spacing. Default value is `DrawingContext.defaultCharacterSpacing`.
+    /// Minimum value is `DrawingContext.minCharacterSpacing`, maximum value is `DrawingContext.maxCharacterSpacing`.
+    public var characterSpacing: Float {
+        get {
+            return HPDF_Page_GetCharSpace(_page)
+        }
+        set {
+            precondition(newValue >= DrawingContext.minCharacterSpacing &&
+                         newValue <= DrawingContext.maxCharacterSpacing,
+                         """
+                         The value of characterSpacing must be between `DrawingContext.minCharacterSpacing` \
+                         and `DrawingContext.maxCharacterSpacing`.
+                         """)
+            if HPDF_Page_SetCharSpace(_page, newValue) != UInt(HPDF_OK) {
+                 HPDF_ResetError(_documentHandle)
+            }
+        }
+    }
+
+    /// The default word spacing.
+    public static let defaultWordSpacing = Float(HPDF_DEF_WORDSPACE)
+
+    /// The minimum word spacing.
+    public static let minWordSpacing = Float(HPDF_MIN_WORDSPACE)
+
+    /// The maximum word spacing.
+    public static let maxWordSpacing = Float(HPDF_MAX_WORDSPACE)
+
+    /// The current value of the page's word spacing. Default value is `DrawingContext.defaultWordSpacing`.
+    /// Minimum value is `DrawingContext.minWordSpacing`, maximum value is `DrawingContext.maxWordSpacing`.
+    public var wordSpacing: Float {
+        get {
+            return HPDF_Page_GetWordSpace(_page)
+        }
+        set {
+            precondition(newValue >= DrawingContext.minWordSpacing && newValue <= DrawingContext.maxWordSpacing,
+                         """
+                         The value of wordSpacing must be between `DrawingContext.minWordSpacing` \
+                         and `DrawingContext.maxWordSpacing`.
+                         """)
+            if HPDF_Page_SetWordSpace(_page, newValue) != UInt(HPDF_OK) {
+                HPDF_ResetError(_documentHandle)
+            }
         }
     }
 
@@ -770,8 +854,6 @@ public final class DrawingContext {
     }
 
     /// Prints the text inside the specified region.
-    ///
-    /// This function does not use the active text matrix.
     ///
     /// - Parameters:
     ///   - text:      The text to show.