+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - Setting text rendering mode
 - Specifying a text matrix when showing text
+- Setting character spacing and word spacing
+- Measuring the number of characters that fit in a width
 - All the enums now conform to `CaseIterable`
 
 ## [0.3.0](https://github.com/WeirdMath/SwiftyHaru/tree/0.3.0) (2018-12-24)